--- processing/src/main/java/org/apache/druid/java/util/metrics/MergeBufferPoolMonitor.java
@@ -1,54 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.java.util.metrics;
-
-import com.google.inject.Inject;
-import org.apache.druid.collections.BlockingPool;
-import org.apache.druid.guice.annotations.Merging;
-import org.apache.druid.java.util.emitter.service.ServiceEmitter;
-import org.apache.druid.java.util.emitter.service.ServiceMetricEvent;
-
-import java.nio.ByteBuffer;
-
-public class MergeBufferPoolMonitor extends AbstractMonitor
-{
-
-  private final BlockingPool<ByteBuffer> mergeBufferPool;
-
-  @Inject
-  public MergeBufferPoolMonitor(
-      @Merging BlockingPool<ByteBuffer> mergeBufferPoolIn
-  )
-  {
-    this.mergeBufferPool = mergeBufferPoolIn;
-  }
-
-  @Override
-  public boolean doMonitor(ServiceEmitter emitter)
-  {
-    long pendingQueries = this.mergeBufferPool.getPendingQueries();
-
-    ServiceMetricEvent.Builder builder = ServiceMetricEvent.builder().setFeed("metrics");
-    emitter.emit(builder.setMetric("mergebuffer/pendingQueries", pendingQueries));
-
-    return true;
-  }
-
-}
